+++ src/main/java/com/ky/docstory/auth/CustomSuccessHandler.java
@@ -1,5 +1,6 @@
 package com.ky.docstory.auth;
 
+import com.ky.docstory.jwt.JWTUtil;
 import jakarta.servlet.http.Cookie;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
